requirements.txt
alembic==1.7.6
click==8.0.3
Flask==2.0.2
flask_bootstrap==3.3.7.1
Flask_Login==0.5.0
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
pycryptodome==3.14.1
SQLAlchemy==1.4.29
Werkzeug==2.0.2
WTForms==3.0.0

pip~=21.3.1
toml~=0.10.2
wheel~=0.37.1
Mako~=1.1.6
MarkupSafe~=2.0.1
yapf~=0.32.0
requests~=2.27.1
dataclasses~=0.8
Jinja2~=3.0.3
itsdangerous~=2.0.1
isort~=5.10.1
pipreqs~=0.4.11
setuptools~=59.6.0
Naked~=0.1.31
yarg~=0.1.9
visitor~=0.1.3
pylint~=2.12.2
astroid~=2.9.3
mccabe~=0.6.1
platformdirs~=2.4.0
wrapt~=1.13.3
certifi~=2021.10.8
docopt~=0.6.2
dominate~=2.6.0
greenlet~=1.1.2
idna~=3.3
urllib3~=1.26.8
zipp~=3.6.0
PyYAML~=6.0
PyMySQL~=1.0.2